More restructuring of folders for easier further development
