+++ src/components/admin/components/global/Header.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../css/App.css'
+import '../../css/App.css'
 
 function Header({ title }) {
     return (
