+++ src/components/admin/pages/Bikes/Bikes.js
@@ -2,7 +2,7 @@ import React from 'react'
 import axios from 'axios'
 
 import Bike from './BikeSingle'
-import Header from '../../global/Header'
+import Header from '../../components/global/Header'
 import DocumentTitle from 'react-document-title'
 
 import { useState, useEffect } from 'react'
