+++ src/components/admin/pages/Home.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import '../css/Other.css'
 import '../css/App.css'
-import Header from '../global/Header'
-import Map from '../Maps/Map'
+import Header from '../components/global/Header'
+import Map from '../components/maps/Map'
 import DocumentTitle from 'react-document-title'
 
 function Home() {
